move file gen methods into separate mixin modules
